+++ examples/egfr-bindcraft/run-dual-gpu.sh
@@ -6,8 +6,6 @@
 
 PIPELINE_DIR=../../
 
-DATESTAMP=$(date +%Y%m%d_%H%M%S)
-
 nextflow run ${PIPELINE_DIR}/main.nf \
   -c nextflow.dual-gpu.config \
   --method bindcraft \
@@ -21,9 +19,7 @@ nextflow run ${PIPELINE_DIR}/main.nf \
   --bindcraft_advanced_settings_preset "default_4stage_multimer" \
   --gpu_devices=0,1 \
   -profile local \
-  -resume \
-  -with-report results/logs/report_${DATESTAMP}.html \
-  -with-trace results/logs/trace_${DATESTAMP}.txt
+  -resume
 
 #  --binder_length_range "65-120" \
 #  --contigs "[F2-23/F84-175/F205-267/0 G91-171/G209-263/0]"