+++ examples/pdl1-bindcraft/run-m3.sh
@@ -2,8 +2,6 @@
 
 PIPELINE_DIR=../../software/nf-binder-design
 
-DATESTAMP=$(date +%Y%m%d_%H%M%S)
-
 DEFAULT_SLURM_ACCOUNT=$(sacctmgr --parsable2 show user -s ${USER} | tail -1 | cut -f 2 -d \|)
 
 nextflow run ${PIPELINE_DIR}/bindcraft.nf \
@@ -18,9 +16,7 @@ nextflow run ${PIPELINE_DIR}/bindcraft.nf \
   --bindcraft_advanced_settings_preset "default_4stage_multimer" \
   # --do_foldseek \
   -profile slurm,m3 \
-  -resume \
-  -with-report results/logs/report_${DATESTAMP}.html \
-  -with-trace results/logs/trace_${DATESTAMP}.txt
+  -resume
 
 # Alternatively, instead of --target_chains you can specify RFDiffusion-style 
 # contigs defining the regions to use, eg: