+++ examples/pdl1-boltzgen/run-local.sh
@@ -2,8 +2,6 @@
 
 PIPELINE_DIR=../../
 
-DATESTAMP=$(date +%Y%m%d_%H%M%S)
-
 nextflow run ${PIPELINE_DIR}/main.nf \
   --method boltzgen \
   --config_yaml pdl1-binder.yaml \
@@ -17,6 +15,4 @@ nextflow run ${PIPELINE_DIR}/main.nf \
   --foldseek_database CATH50 \
   --foldseek_databases_path "$(pwd)/../databases/foldseek" \
   -profile local \
-  -resume \
-  -with-report results/logs/report_${DATESTAMP}.html \
-  -with-trace results/logs/trace_${DATESTAMP}.txt
+  -resume