+++ examples/pdl1-rfd/run-local-simple.sh
@@ -2,8 +2,6 @@
 
 PIPELINE_DIR=../../
 
-DATESTAMP=$(date +%Y%m%d_%H%M%S)
-
 nextflow run ${PIPELINE_DIR}/main.nf \
   --method rfd \
   --input_pdb 'input/*.pdb' \
@@ -12,6 +10,4 @@ nextflow run ${PIPELINE_DIR}/main.nf \
   --hotspot_res "A56" \
   --rfd_n_designs=4 \
   -profile local \
-  -resume \
-  -with-report results/logs/report_${DATESTAMP}.html \
-  -with-trace results/logs/trace_${DATESTAMP}.txt
+  -resume